+++ pymsis/__init__.py
@@ -3,8 +3,9 @@
 import importlib.metadata
 
 from pymsis.msis import Variable, calculate
+from pymsis.utils import use_space_weather_file
 
 
 __version__ = importlib.metadata.version("pymsis")
 
-__all__ = ["Variable", "__version__", "calculate"]
+__all__ = ["Variable", "__version__", "calculate", "use_space_weather_file"]