+++ pymsis/utils.py
@@ -1,5 +1,6 @@
 """Utilities for obtaining input datasets."""
 
+import os
 import urllib.request
 import warnings
 from io import BytesIO
@@ -13,8 +14,50 @@ import pymsis
 
 _DATA_FNAME: str = "SW-All.csv"
 _F107_AP_URL: str = f"https://celestrak.org/SpaceData/{_DATA_FNAME}"
-_F107_AP_PATH: Path = Path(pymsis.__file__).parent / _DATA_FNAME
+_F107_AP_DEFAULT_FILE: Path = Path(pymsis.__file__).parent / _DATA_FNAME
 _DATA: dict[str, npt.NDArray] | None = None
+
+_F107_AP_FILE: Path = Path(
+    os.environ.get("PYMSIS_SPACE_WEATHER_FILE", _F107_AP_DEFAULT_FILE)
+)
+
+
+def use_space_weather_file(file: str | Path | None = None) -> None:
+    """
+    Direct pymsis to use a custom file to retrieve the F10.7 and ap data from.
+
+    The custom data file must follow the same (.csv) format as data retrieved from
+    Celestrak. The legacy (.txt) file format is not supported.
+
+    By default, the data is retrieved from CelesTrak and stored inside the
+    installed package location. Retrieving the data from a custom file path may
+    be useful if you are retrieving the data from a different source, if you
+    have a centralized location for the data, or would like to use custom data.
+
+    Alternatively, the file path can be set using the environment variable
+    ``PYMSIS_SPACE_WEATHER_FILE`` to achieve the same result and
+    to avoid setting the space weather file path programmatically on each run.
+
+    Setting a default and running `download_f107_ap()` will not download to
+    this custom file path.
+
+    Parameters
+    ----------
+    file : str or Path or None
+        Path to the F10.7 and ap data file, retrieved from CelesTrak or elsewhere.
+        If set to None, the space weather file downloaded from CelesTrak (stored in the
+        installed package location) will be used.
+    """
+    file = Path(file) if file is not None else _F107_AP_DEFAULT_FILE
+    if not file.is_file():
+        raise FileNotFoundError(
+            f"Provided custom space weather file does not exist: {file}"
+        )
+
+    # update the global path and data variables
+    global _F107_AP_FILE, _DATA  # noqa: PLW0603
+    _F107_AP_FILE = Path(file)
+    _DATA = None
 
 
 def download_f107_ap() -> None:
@@ -25,6 +68,9 @@ def download_f107_ap() -> None:
     the same filename as the data source: ``SW-All.csv``.
     This routine can be called to update the data as well if you would like to
     use newer data since the last time you downloaded the file.
+
+    If `use_space_weather_file()` has been called to set a custom file path, the file
+    will still be downloaded to the default location and thus ignored by pymsis.
 
     Notes
     -----
@@ -43,14 +89,30 @@ def download_f107_ap() -> None:
        Space Weather, https://doi.org/10.1029/2020SW002641
     """
     warnings.warn(f"Downloading ap and F10.7 data from {_F107_AP_URL}")
+    if _F107_AP_DEFAULT_FILE != _F107_AP_FILE:
+        warnings.warn(
+            "A custom space weather file has been set, but the downloaded file "
+            "will be stored in the default location and ignored. Unset the "
+            "custom file path using `use_space_weather_file(None)` to use the "
+            "downloaded file."
+        )
     req = urllib.request.urlopen(_F107_AP_URL)
-    with _F107_AP_PATH.open("wb") as f:
+    with _F107_AP_DEFAULT_FILE.open("wb") as f:
         f.write(req.read())
 
 
 def _load_f107_ap_data() -> dict[str, npt.NDArray]:
     """Load data from disk, if it isn't present go out and download it first."""
-    if not _F107_AP_PATH.exists():
+    default_file_exists = _F107_AP_DEFAULT_FILE.is_file()
+    custom_file_used = _F107_AP_FILE != _F107_AP_DEFAULT_FILE
+
+    if custom_file_used and not _F107_AP_FILE.is_file():
+        raise FileNotFoundError(
+            "Custom space weather file has been set but does not exist: "
+            f"{_F107_AP_FILE}"
+        )
+
+    if not custom_file_used and not default_file_exists:
         download_f107_ap()
 
     dtype = {
@@ -90,7 +152,7 @@ def _load_f107_ap_data() -> dict[str, npt.NDArray]:
     # Use a buffer to read in and load so we can quickly get rid of
     # the extra "PRD" lines at the end of the file (unknown length
     # so we can't just go back in line lengths)
-    with _F107_AP_PATH.open() as fin:
+    with _F107_AP_FILE.open() as fin:
         with BytesIO() as fout:
             for line in fin:
                 if "PRM" in line:
@@ -168,7 +230,8 @@ def _load_f107_ap_data() -> dict[str, npt.NDArray]:
         "f107a": f107a_data,
         "warn_data": warn_data,
     }
-    globals()["_DATA"] = data
+    global _DATA  # noqa: PLW0603
+    _DATA = data
     return data